chore: update model settings
