+++ web-app/src/lib/predefined.ts
@@ -1,223 +1,299 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-export const presetConfiguration: Record<string, any> = {
-  prompt_template: {
-    key: 'prompt_template',
-    title: 'Prompt template',
-    description: `A structured format that guides how the model should respond.`,
-    controller_type: 'textarea',
+export const modelSettings = {
+  ctx_size: {
+    key: 'ctx_size',
+    title: 'Context Size',
+    description: 'Size of the prompt context (0 = loaded from model).',
+    controller_type: 'input',
     controller_props: {
-      placeholder: 'Prompt template',
-      value: '',
-      rows: 4,
+      value: 8192,
+      placeholder: '8192',
+      type: 'number',
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  stop: {
-    key: 'stop',
-    title: 'Stop',
-    description: `Defines tokens or phrases that will end the model's response.`,
-    controller_type: 'tag',
+  n_predict: {
+    key: 'n_predict',
+    title: 'Max Tokens to Predict',
+    description: 'Maximum number of tokens to generate (-1 = infinity).',
+    controller_type: 'input',
     controller_props: {
-      placeholder: 'Enter stop words',
-      value: [''],
+      value: -1,
+      placeholder: '-1',
+      type: 'number',
     },
-    require_model_reload: false,
-    config_type: 'runtime',
   },
-  ctx_len: {
-    key: 'ctx_len',
-    title: 'Context Length',
-    description: `Controls how much text the model can consider at once. Longer context allows the model to handle more input but uses more memory and runs slower.`,
-    controller_type: 'slider',
+  batch_size: {
+    key: 'batch_size',
+    title: 'Batch Size',
+    description: 'Logical maximum batch size for processing prompts.',
+    controller_type: 'input',
     controller_props: {
-      min: 128,
-      max: 4096,
-      step: 128,
-      value: 4096,
+      value: 2048,
+      placeholder: '2048',
+      type: 'number',
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  max_tokens: {
-    key: 'max_tokens',
-    title: 'Max Tokens',
-    description: `Controls response length. Higher values allow longer, more detailed responses.`,
-    controller_type: 'slider',
+  ubatch_size: {
+    key: 'ubatch_size',
+    title: 'uBatch Size',
+    description: 'Physical maximum batch size for processing prompts.',
+    controller_type: 'input',
     controller_props: {
-      min: 100,
-      max: 4096,
-      step: 10,
-      value: 4096,
+      value: 512,
+      placeholder: '512',
+      type: 'number',
     },
-    require_model_reload: false,
-    config_type: 'runtime',
   },
-  ngl: {
-    key: 'ngl',
-    title: 'Number of GPU layers (ngl)',
+  n_gpu_layers: {
+    key: 'n_gpu_layers',
+    title: 'GPU Layers',
     description:
-      'Controls how many layers of the model run on GPU. More layers means faster processing, but requires more GPU memory.',
-    controller_type: 'slider',
+      'Number of model layers to offload to the GPU (-1 for all layers, 0 for CPU only).',
+    controller_type: 'input',
     controller_props: {
-      min: 1,
-      max: 100,
-      step: 1,
-      value: 100,
+      value: -1,
+      placeholder: '-1',
+      type: 'number',
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  embedding: {
-    key: 'embedding',
-    title: 'Embedding',
-    description: 'Whether to enable embedding.',
-    controller_type: 'checkbox',
+  device: {
+    key: 'device',
+    title: 'Devices for Offload',
+    description:
+      "Comma-separated list of devices to use for offloading (e.g., 'cuda:0', 'cuda:0,cuda:1'). Leave empty to use default/CPU only.",
+    controller_type: 'input',
     controller_props: {
-      value: true,
+      value: '',
+      placeholder: 'cuda:0',
+      type: 'text',
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  stream: {
-    key: 'stream',
-    title: 'Stream',
-    description: `Enables real-time response streaming.`,
+  split_mode: {
+    key: 'split_mode',
+    title: 'GPU Split Mode',
+    description: 'How to split the model across multiple GPUs.',
+    controller_type: 'dropdown',
+    controller_props: {
+      value: 'layer',
+      options: [
+        { value: 'none', name: 'None' },
+        { value: 'layer', name: 'Layer' },
+        { value: 'row', name: 'Row' },
+      ],
+    },
+  },
+  main_gpu: {
+    key: 'main_gpu',
+    title: 'Main GPU Index',
+    description:
+      'The GPU to use for the model (split-mode=none) or intermediate results (split-mode=row).',
+    controller_type: 'input',
+    controller_props: {
+      value: 0,
+      placeholder: '0',
+      type: 'number',
+    },
+  },
+  flash_attn: {
+    key: 'flash_attn',
+    title: 'Flash Attention',
+    description: 'Enable Flash Attention for optimized performance.',
     controller_type: 'checkbox',
     controller_props: {
       value: false,
     },
-    require_model_reload: false,
-    config_type: 'runtime',
   },
-  temperature: {
-    key: 'temperature',
-    title: 'Temperature',
-    description: `Controls response randomness. Higher values produce more creative, varied responses. `,
-    controller_type: 'slider',
-    controller_props: {
-      min: 0,
-      max: 2,
-      step: 0.1,
-      value: 0.7,
-    },
-    require_model_reload: false,
-    config_type: 'runtime',
-  },
-  frequency_penalty: {
-    key: 'frequency_penalty',
-    title: 'Frequency Penalty',
-    description: `Reduces word repetition. Higher values encourage more varied language. Useful for creative writing and content generation.`,
-    controller_type: 'slider',
-    controller_props: {
-      min: 0,
-      max: 1,
-      step: 0.1,
-      value: 0.7,
-    },
-    require_model_reload: false,
-    config_type: 'runtime',
-  },
-  presence_penalty: {
-    key: 'presence_penalty',
-    title: 'Presence Penalty',
-    description: `Encourages the model to explore new topics. Higher values help prevent the model from fixating on already-discussed subjects.`,
-    controller_type: 'slider',
-    controller_props: {
-      min: 0,
-      max: 1,
-      step: 0.1,
-      value: 0.7,
-    },
-    require_model_reload: false,
-    config_type: 'runtime',
-  },
-  top_p: {
-    key: 'top_p',
-    title: 'Top P',
-    description: `Set probability threshold for more relevant outputs. Higher values allow more diverse word choices.`,
-    controller_type: 'slider',
-    controller_props: {
-      min: 0,
-      max: 1,
-      step: 0.1,
-      value: 0.95,
-    },
-    require_model_reload: false,
-    config_type: 'runtime',
-  },
-  cpu_threads: {
-    key: 'cpu_threads',
-    title: 'CPU Threads',
+  cont_batching: {
+    key: 'cont_batching',
+    title: 'Continuous Batching',
     description:
-      'Controls how many CPU cores are used for processing. Higher values can speed up model loading but use more system resources.',
-    controller_type: 'slider',
+      'Enable continuous batching (a.k.a dynamic batching) for concurrent requests (default: enabled).',
+    controller_type: 'checkbox',
     controller_props: {
+      value: true,
+    },
+  },
+  no_mmap: {
+    key: 'no_mmap',
+    title: 'Disable mmap',
+    description:
+      'Do not memory-map model (slower load but may reduce pageouts if not using mlock).',
+    controller_type: 'checkbox',
+    controller_props: {
+      value: false,
+    },
+  },
+  mlock: {
+    key: 'mlock',
+    title: 'MLock',
+    description:
+      'Force system to keep model in RAM, preventing swapping/compression.',
+    controller_type: 'checkbox',
+    controller_props: {
+      value: false,
+    },
+  },
+  no_kv_offload: {
+    key: 'no_kv_offload',
+    title: 'Disable KV Offload',
+    description: 'Disable KV cache offload to GPU (if GPU is used).',
+    controller_type: 'checkbox',
+    controller_props: {
+      value: false,
+    },
+  },
+  cache_type_k: {
+    key: 'cache_type_k',
+    title: 'KV Cache K Type',
+    description: 'KV cache data type for Keys (default: f16).',
+    controller_type: 'dropdown',
+    controller_props: {
+      value: 'f16',
+      options: [
+        { value: 'f32', name: 'f32' },
+        { value: 'f16', name: 'f16' },
+        { value: 'bf16', name: 'bf16' },
+        { value: 'q8_0', name: 'q8_0' },
+        { value: 'q4_0', name: 'q4_0' },
+        { value: 'q4_1', name: 'q4_1' },
+        { value: 'iq4_nl', name: 'iq4_nl' },
+        { value: 'q5_0', name: 'q5_0' },
+        { value: 'q5_1', name: 'q5_1' },
+      ],
+    },
+  },
+  cache_type_v: {
+    key: 'cache_type_v',
+    title: 'KV Cache V Type',
+    description: 'KV cache data type for Values (default: f16).',
+    controller_type: 'dropdown',
+    controller_props: {
+      value: 'f16',
+      options: [
+        { value: 'f32', name: 'f32' },
+        { value: 'f16', name: 'f16' },
+        { value: 'bf16', name: 'bf16' },
+        { value: 'q8_0', name: 'q8_0' },
+        { value: 'q4_0', name: 'q4_0' },
+        { value: 'q4_1', name: 'q4_1' },
+        { value: 'iq4_nl', name: 'iq4_nl' },
+        { value: 'q5_0', name: 'q5_0' },
+        { value: 'q5_1', name: 'q5_1' },
+      ],
+    },
+  },
+  defrag_thold: {
+    key: 'defrag_thold',
+    title: 'KV Cache Defragmentation Threshold',
+    description: 'Threshold for KV cache defragmentation (< 0 to disable).',
+    controller_type: 'input',
+    controller_props: {
+      value: 0.1,
+      placeholder: '0.1',
+      type: 'number',
+      step: 0.01,
+    },
+  },
+  rope_scaling: {
+    key: 'rope_scaling',
+    title: 'RoPE Scaling Method',
+    description: 'RoPE frequency scaling method.',
+    controller_type: 'dropdown',
+    controller_props: {
+      value: 'none',
+      options: [
+        { value: 'none', name: 'None' },
+        { value: 'linear', name: 'Linear' },
+        { value: 'yarn', name: 'YaRN' },
+      ],
+    },
+  },
+  rope_scale: {
+    key: 'rope_scale',
+    title: 'RoPE Scale Factor',
+    description: 'RoPE context scaling factor.',
+    controller_type: 'input',
+    controller_props: {
+      value: 1.0,
+      placeholder: '1.0',
+      type: 'number',
       min: 0,
-      max: 128,
-      step: 1,
-      value: 1,
+      step: 0.01,
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  // assistant
-  chunk_size: {
-    key: 'chunk_size',
-    title: 'Chunk Size',
+  rope_freq_base: {
+    key: 'rope_freq_base',
+    title: 'RoPE Frequency Base',
+    description: 'RoPE base frequency (0 = loaded from model).',
+    controller_type: 'input',
+    controller_props: {
+      value: 0,
+      placeholder: '0 (model default)',
+      type: 'number',
+    },
+  },
+  rope_freq_scale: {
+    key: 'rope_freq_scale',
+    title: 'RoPE Frequency Scale Factor',
+    description: 'RoPE frequency scaling factor.',
+    controller_type: 'input',
+    controller_props: {
+      value: 1.0,
+      placeholder: '1.0',
+      type: 'number',
+      min: 0,
+      step: 0.01,
+    },
+  },
+  mirostat: {
+    key: 'mirostat',
+    title: 'Mirostat Mode',
     description:
-      'Sets the maximum number of tokens per data chunk. Increase size for processing large blocks of text efficiently, or decrease it to optimize memory usage.',
-    controller_type: 'slider',
+      'Use Mirostat sampling (0: disabled, 1: Mirostat V1, 2: Mirostat V2).',
+    controller_type: 'dropdown',
     controller_props: {
-      min: 128,
-      max: 2048,
-      step: 128,
-      value: 1024,
+      value: 0,
+      options: [
+        { value: 0, name: 'Disabled' },
+        { value: 1, name: 'Mirostat V1' },
+        { value: 2, name: 'Mirostat V2' },
+      ],
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  chunk_overlap: {
-    key: 'chunk_overlap',
-    title: 'Chunk Overlap',
-    description:
-      'How many words overlap between text segments. More overlap helps maintain meaning across splits.',
-    controller_type: 'slider',
+  mirostat_lr: {
+    key: 'mirostat_lr',
+    title: 'Mirostat Learning Rate',
+    description: 'Mirostat learning rate (eta).',
+    controller_type: 'input',
     controller_props: {
-      min: 32,
-      max: 512,
-      step: 32,
-      value: 64,
+      value: 0.1,
+      placeholder: '0.1',
+      type: 'number',
+      min: 0,
+      step: 0.01,
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
-  top_k: {
-    key: 'top_k',
-    title: 'Top K',
-    description:
-      'Number of most relevant documents to retrieve. Higher values return more results.',
-    controller_type: 'slider',
+  mirostat_ent: {
+    key: 'mirostat_ent',
+    title: 'Mirostat Target Entropy',
+    description: 'Mirostat target entropy (tau).',
+    controller_type: 'input',
     controller_props: {
-      min: 1,
-      max: 5,
-      step: 1,
-      value: 2,
+      value: 5.0,
+      placeholder: '5.0',
+      type: 'number',
+      min: 0,
+      step: 0.01,
     },
-    require_model_reload: false,
-    config_type: 'runtime',
   },
-  retrieval_template: {
-    key: 'retrieval_template',
-    title: 'Retrieval Template',
-    description: 'Format for how the AI uses context to answer questions.',
-    controller_type: 'textarea',
+  json_schema_file: {
+    key: 'json_schema_file',
+    title: 'JSON Schema File',
+    description: 'Path to a JSON schema file to constrain generations.',
+    controller_type: 'input',
     controller_props: {
-      placeholder: 'Retrieval Template',
       value: '',
-      rows: 4,
+      placeholder: 'path/to/schema.json',
+      type: 'text',
     },
-    require_model_reload: true,
-    config_type: 'setting',
   },
 }
